+++ src/module/Process.hpp
@@ -16,5 +16,7 @@ namespace be {
         static JSValue setTimezone(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
         static JSValue readEFUSE(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
         static JSValue readMac(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        static JSValue gc(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        static JSValue ref(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
     } ;
 }